nvim: Draft (and unfinished) lua implementation of clang-format.py
I'm currently using clang-format from null-ls but the range formatting
is not working properly so this was an early attempt to my own
implementation. I'm committing this just in case but I would rather
figure out what is going on with the default one.

It could be the actual invocation of clang-format in null-ls or the lsp
mapping that I've set up.