ssh: configuration and utility functions
Add encrypted file for remote "home" hosts. Also add the `ssh` wrapper
which will open for hosts that have confiig in `remote_machine` file an
ssh session in a new window outsize of the local tmux session. This is
because usually the remote machine that I use have their own tmux
session.